cleanup vis event interface
This removes the function pointer interface which was adding
needless complexity and making it difficult to add new events. Now
if new events are only meant for lua they only need to be added to
the lua interface. This will also have a minor reduction in
runtime memory usage and produce a smaller binary.

The only runtime difference is that QUIT happens after all windows
have been closed and their files freed.
